requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
duckdb==1.1.3
pandas==2.2.2
scikit-learn==1.5.2
joblib==1.4.2
typer==0.12.5
streamlit==1.38.0
requests==2.32.3
pydantic-settings==2.5.2
pyyaml==6.0.2
